harness/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ie-agent-harness"
version = "0.1.0"
description = "Standalone buyer-agent runner for the Information Exchange marketplace"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.87.0",
    "pyyaml>=6.0.2",
    "requests>=2.33.0",
]
